Import of GFUJI/macro-0.02 from CPAN.
gitpan-cpan-distribution: macro
gitpan-cpan-version:      0.02
gitpan-cpan-path:         GFUJI/macro-0.02.tar.gz
gitpan-cpan-author:       GFUJI
gitpan-cpan-maturity:     released

+++ t/05_pmc.t
@@ -1,11 +1,14 @@
 #!perl -w
 use strict;
 
-use Test::More tests => 9;
+use Test::More tests => 13;
+
+my $debug = $ENV{PERL_MACRO_DEBUG};
 
 use FindBin qw($Bin);
 use lib "$Bin/tlib";
 use Fatal qw(open unlink);
+
 
 my $pm1 = "$Bin/tlib/Foo.pm";
 my $pm2 = "$Bin/tlib/Bar.pm";
@@ -18,7 +21,8 @@ unlink($pm2.'c') if -e $pm2.'c';
 	use File::Spec;
 
 	open my $save_stderr, '>&', \*STDERR;
-	open *STDERR, '>', File::Spec->devnull;
+	open *STDERR, '>', File::Spec->devnull
+		unless $debug;
 
 	is system($^X, '-c', "-I$Bin/../lib", $pm1), 0, 'compile Foo.pm';
 	is system($^X, '-c', "-I$Bin/../lib", $pm2), 0, 'compile Bar.pm';
@@ -37,8 +41,20 @@ is Foo::g(), 'Foo::g', 'Foo::g()';
 is Bar::f(), 'Bar::f', 'Bar::f()';
 is Bar::g(), 'Bar::g', 'Bar::g()';
 
+
+is Foo::h(), 'func', 'lexicality in Foo';
+is Bar::h(), 'func', 'lexicality in Bar';
+
+{
+	local $TODO = 'Line adjustment not yet implemented';
+	is Foo::line(), Foo::correct_line(), 'Foo: correct lineno';
+	is Bar::line(), Bar::correct_line(), 'Bar: correct lineno';
+}
+
 ok !$INC{'macro.pm'}, 'macro.pm was not loaded';
 
-unlink($pm1.'c');
-unlink($pm2.'c');
+unless($debug){
+	unlink($pm1.'c');
+	unlink($pm2.'c');
+}
